moved projectile classes into Projectiles package in Actors and updated effected classes
