+++ src/main/java/com/example/demo/Actors/Projectiles/EnemyProjectile.java
@@ -1,4 +1,4 @@
-package com.example.demo.Actors;
+package com.example.demo.Actors.Projectiles;
 
 public class EnemyProjectile extends Projectile {
